requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
python-dotenv==0.19.2
pydantic==2.0.3
PyMySQL==1.1.0
starlette==0.27.0
typing_extensions==4.7.1
